# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Renta de Apartamentos - API",
        "description": "Documentación de la API del sistema de rentas de apartamentos. Permite gestionar usuarios, apartamentos, quejas, contratos y más.",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "license": {
            "name": "Todos los derechos reservados - Universidad Piloto de Colombia"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Listar todos los usuarios",
                "operationId": "e9d4d8b419dbef04b5748fbf9ccf3881",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios obtenida correctamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "4e9dbcbdaaaf9faab196a82d5cbaea23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "correo",
                                    "telefono",
                                    "contraseña"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3012345678"
                                    },
                                    "contraseña": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secreto123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
...
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Mostrar información de un usuario específico",
                "operationId": "e9ac05ddc4ea46d8848b012d7d92e3ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar un usuario existente",
                "operationId": "27ac608b6dd8b8e7c7c45b10c5210b3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "correo",
                                    "telefono"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3012345678"
                                    },
                                    "contraseña": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "nuevaClave123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "ea700764859a665afd38bb39c2842f1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/api/usuarios/{id}/edit": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Obtener datos para editar un usuario",
                "operationId": "f340e93d3b00218a2e8fb92d59c96d61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del usuario para edición",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Usuario": {
                "title": "Usuario",
                "required": [
                    "nombre",
                    "correo",
                    "telefono",
                    "contraseña"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "123456789"
                    },
                    "contraseña": {
                        "type": "string",
                        "example": "********"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}